lgtm warning

+++ aws-android-sdk-s3/src/main/java/com/amazonaws/mobileconnectors/s3/transferutility/UploadPartTask.java
@@ -61,7 +61,7 @@ class UploadPartTask implements Callable<Boolean> {
         while (true) {
             try {
                 final UploadPartResult putPartResult = s3.uploadPart(uploadPartRequest);
-                 setTaskState(TransferState.PART_COMPLETED);
+                setTaskState(TransferState.PART_COMPLETED);
                 dbUtil.updateETag(uploadPartRequest.getId(), putPartResult.getETag());
                 return true;
             } catch (AbortedException e) {
@@ -149,8 +149,8 @@ class UploadPartTask implements Callable<Boolean> {
     }
 
     private long exponentialBackoffWithJitter(int retryAttempt) {
-        final int baseTimeMs = 1000;
-        final int jitterFactor = 1000;
+        final long baseTimeMs = 1000L;
+        final long jitterFactor = 1000L;
         long delay = baseTimeMs * (1 << retryAttempt);
         long jitter = (long) (jitterFactor * Math.random());
         return delay + jitter;
